backend/pyproject.toml
[project]
name = "kds-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.32",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.6.6",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "requests-html>=0.10.0",
    "sentence-transformers>=5.1.0",
]
